+++ crdesigner/crmapver/verification/hol/parser/gen/FormulaLexer.py
@@ -1,0 +1,213 @@
+# Generated from Formula.g4 by ANTLR 4.9.3
+from antlr4 import *
+from io import StringIO
+import sys
+if sys.version_info[1] > 5:
+    from typing import TextIO
+else:
+    from typing.io import TextIO
+
+
+
+def serializedATN():
+    with StringIO() as buf:
+        buf.write("\3\u608b\ua72a\u8133\ub9ed\u417c\u3be7\u7786\u5964\2!")
+        buf.write("\u011b\b\1\4\2\t\2\4\3\t\3\4\4\t\4\4\5\t\5\4\6\t\6\4\7")
+        buf.write("\t\7\4\b\t\b\4\t\t\t\4\n\t\n\4\13\t\13\4\f\t\f\4\r\t\r")
+        buf.write("\4\16\t\16\4\17\t\17\4\20\t\20\4\21\t\21\4\22\t\22\4\23")
+        buf.write("\t\23\4\24\t\24\4\25\t\25\4\26\t\26\4\27\t\27\4\30\t\30")
+        buf.write("\4\31\t\31\4\32\t\32\4\33\t\33\4\34\t\34\4\35\t\35\4\36")
+        buf.write("\t\36\4\37\t\37\4 \t \4!\t!\4\"\t\"\3\2\3\2\3\3\3\3\3")
+        buf.write("\4\3\4\3\5\3\5\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\5\6V\n")
+        buf.write("\6\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\5\7b\n\7\3")
+        buf.write("\b\3\b\3\b\3\b\3\b\3\b\3\b\5\bk\n\b\3\t\3\t\3\t\3\t\5")
+        buf.write("\tq\n\t\3\n\3\n\3\n\3\n\3\n\3\n\3\n\5\nz\n\n\3\13\3\13")
+        buf.write("\3\13\3\13\3\13\3\13\5\13\u0082\n\13\3\f\3\f\3\f\3\f\3")
+        buf.write("\f\3\f\3\f\5\f\u008b\n\f\3\r\3\r\3\r\3\r\3\r\5\r\u0092")
+        buf.write("\n\r\3\16\3\16\3\16\3\16\3\16\3\16\3\16\5\16\u009b\n\16")
+        buf.write("\3\17\3\17\3\17\3\20\3\20\3\20\3\20\3\21\3\21\3\21\3\22")
+        buf.write("\3\22\3\23\3\23\3\23\3\24\3\24\3\25\3\25\3\26\3\26\3\26")
+        buf.write("\3\27\3\27\3\27\3\30\3\30\3\30\3\31\6\31\u00ba\n\31\r")
+        buf.write("\31\16\31\u00bb\3\31\5\31\u00bf\n\31\3\31\6\31\u00c2\n")
+        buf.write("\31\r\31\16\31\u00c3\7\31\u00c6\n\31\f\31\16\31\u00c9")
+        buf.write("\13\31\3\32\6\32\u00cc\n\32\r\32\16\32\u00cd\3\32\5\32")
+        buf.write("\u00d1\n\32\3\32\6\32\u00d4\n\32\r\32\16\32\u00d5\7\32")
+        buf.write("\u00d8\n\32\f\32\16\32\u00db\13\32\3\33\3\33\3\33\3\34")
+        buf.write("\3\34\3\35\3\35\3\35\3\36\6\36\u00e6\n\36\r\36\16\36\u00e7")
+        buf.write("\3\36\7\36\u00eb\n\36\f\36\16\36\u00ee\13\36\3\36\7\36")
+        buf.write("\u00f1\n\36\f\36\16\36\u00f4\13\36\3\37\6\37\u00f7\n\37")
+        buf.write("\r\37\16\37\u00f8\3\37\3\37\3 \3 \3 \5 \u0100\n \3 \6")
+        buf.write(" \u0103\n \r \16 \u0104\7 \u0107\n \f \16 \u010a\13 \3")
+        buf.write(" \3 \3!\3!\3!\6!\u0111\n!\r!\16!\u0112\5!\u0115\n!\3\"")
+        buf.write("\3\"\3\"\5\"\u011a\n\"\2\2#\3\3\5\4\7\5\t\6\13\7\r\b\17")
+        buf.write("\t\21\n\23\13\25\f\27\r\31\16\33\17\35\20\37\21!\22#\23")
+        buf.write("%\24\'\25)\26+\27-\30/\31\61\2\63\2\65\32\67\339\34;\35")
+        buf.write("=\36?\37A C!\3\2\t\3\2c|\3\2aa\5\2\62;C\\c|\3\2C\\\3\2")
+        buf.write("\62;\4\2\13\f\"\"\3\2\63;\2\u0137\2\3\3\2\2\2\2\5\3\2")
+        buf.write("\2\2\2\7\3\2\2\2\2\t\3\2\2\2\2\13\3\2\2\2\2\r\3\2\2\2")
+        buf.write("\2\17\3\2\2\2\2\21\3\2\2\2\2\23\3\2\2\2\2\25\3\2\2\2\2")
+        buf.write("\27\3\2\2\2\2\31\3\2\2\2\2\33\3\2\2\2\2\35\3\2\2\2\2\37")
+        buf.write("\3\2\2\2\2!\3\2\2\2\2#\3\2\2\2\2%\3\2\2\2\2\'\3\2\2\2")
+        buf.write("\2)\3\2\2\2\2+\3\2\2\2\2-\3\2\2\2\2/\3\2\2\2\2\65\3\2")
+        buf.write("\2\2\2\67\3\2\2\2\29\3\2\2\2\2;\3\2\2\2\2=\3\2\2\2\2?")
+        buf.write("\3\2\2\2\2A\3\2\2\2\2C\3\2\2\2\3E\3\2\2\2\5G\3\2\2\2\7")
+        buf.write("I\3\2\2\2\tK\3\2\2\2\13U\3\2\2\2\ra\3\2\2\2\17j\3\2\2")
+        buf.write("\2\21p\3\2\2\2\23y\3\2\2\2\25\u0081\3\2\2\2\27\u008a\3")
+        buf.write("\2\2\2\31\u0091\3\2\2\2\33\u009a\3\2\2\2\35\u009c\3\2")
+        buf.write("\2\2\37\u009f\3\2\2\2!\u00a3\3\2\2\2#\u00a6\3\2\2\2%\u00a8")
+        buf.write("\3\2\2\2\'\u00ab\3\2\2\2)\u00ad\3\2\2\2+\u00af\3\2\2\2")
+        buf.write("-\u00b2\3\2\2\2/\u00b5\3\2\2\2\61\u00b9\3\2\2\2\63\u00cb")
+        buf.write("\3\2\2\2\65\u00dc\3\2\2\2\67\u00df\3\2\2\29\u00e1\3\2")
+        buf.write("\2\2;\u00e5\3\2\2\2=\u00f6\3\2\2\2?\u00fc\3\2\2\2A\u0114")
+        buf.write("\3\2\2\2C\u0116\3\2\2\2EF\7\60\2\2F\4\3\2\2\2GH\7.\2\2")
+        buf.write("H\6\3\2\2\2IJ\7+\2\2J\b\3\2\2\2KL\7*\2\2L\n\3\2\2\2MN")
+        buf.write("\7V\2\2NO\7t\2\2OP\7w\2\2PV\7g\2\2QR\7v\2\2RS\7t\2\2S")
+        buf.write("T\7w\2\2TV\7g\2\2UM\3\2\2\2UQ\3\2\2\2V\f\3\2\2\2WX\7H")
+        buf.write("\2\2XY\7c\2\2YZ\7n\2\2Z[\7u\2\2[b\7g\2\2\\]\7h\2\2]^\7")
+        buf.write("c\2\2^_\7n\2\2_`\7u\2\2`b\7g\2\2aW\3\2\2\2a\\\3\2\2\2")
+        buf.write("b\16\3\2\2\2ck\7#\2\2de\7p\2\2ef\7q\2\2fk\7v\2\2gh\7P")
+        buf.write("\2\2hi\7Q\2\2ik\7V\2\2jc\3\2\2\2jd\3\2\2\2jg\3\2\2\2k")
+        buf.write("\20\3\2\2\2lq\7C\2\2mn\7C\2\2no\7N\2\2oq\7N\2\2pl\3\2")
+        buf.write("\2\2pm\3\2\2\2q\22\3\2\2\2rz\7G\2\2st\7G\2\2tu\7Z\2\2")
+        buf.write("uv\7K\2\2vw\7U\2\2wx\7V\2\2xz\7U\2\2yr\3\2\2\2ys\3\2\2")
+        buf.write("\2z\24\3\2\2\2{\u0082\7E\2\2|}\7E\2\2}~\7Q\2\2~\177\7")
+        buf.write("W\2\2\177\u0080\7P\2\2\u0080\u0082\7V\2\2\u0081{\3\2\2")
+        buf.write("\2\u0081|\3\2\2\2\u0082\26\3\2\2\2\u0083\u008b\7(\2\2")
+        buf.write("\u0084\u0085\7c\2\2\u0085\u0086\7p\2\2\u0086\u008b\7f")
+        buf.write("\2\2\u0087\u0088\7C\2\2\u0088\u0089\7P\2\2\u0089\u008b")
+        buf.write("\7F\2\2\u008a\u0083\3\2\2\2\u008a\u0084\3\2\2\2\u008a")
+        buf.write("\u0087\3\2\2\2\u008b\30\3\2\2\2\u008c\u0092\7~\2\2\u008d")
+        buf.write("\u008e\7q\2\2\u008e\u0092\7t\2\2\u008f\u0090\7Q\2\2\u0090")
+        buf.write("\u0092\7T\2\2\u0091\u008c\3\2\2\2\u0091\u008d\3\2\2\2")
+        buf.write("\u0091\u008f\3\2\2\2\u0092\32\3\2\2\2\u0093\u009b\7`\2")
+        buf.write("\2\u0094\u0095\7z\2\2\u0095\u0096\7q\2\2\u0096\u009b\7")
+        buf.write("t\2\2\u0097\u0098\7Z\2\2\u0098\u0099\7Q\2\2\u0099\u009b")
+        buf.write("\7T\2\2\u009a\u0093\3\2\2\2\u009a\u0094\3\2\2\2\u009a")
+        buf.write("\u0097\3\2\2\2\u009b\34\3\2\2\2\u009c\u009d\7/\2\2\u009d")
+        buf.write("\u009e\7@\2\2\u009e\36\3\2\2\2\u009f\u00a0\7>\2\2\u00a0")
+        buf.write("\u00a1\7/\2\2\u00a1\u00a2\7@\2\2\u00a2 \3\2\2\2\u00a3")
+        buf.write("\u00a4\7k\2\2\u00a4\u00a5\7p\2\2\u00a5\"\3\2\2\2\u00a6")
+        buf.write("\u00a7\7?\2\2\u00a7$\3\2\2\2\u00a8\u00a9\7#\2\2\u00a9")
+        buf.write("\u00aa\7?\2\2\u00aa&\3\2\2\2\u00ab\u00ac\7>\2\2\u00ac")
+        buf.write("(\3\2\2\2\u00ad\u00ae\7@\2\2\u00ae*\3\2\2\2\u00af\u00b0")
+        buf.write("\7>\2\2\u00b0\u00b1\7?\2\2\u00b1,\3\2\2\2\u00b2\u00b3")
+        buf.write("\7@\2\2\u00b3\u00b4\7?\2\2\u00b4.\3\2\2\2\u00b5\u00b6")
+        buf.write("\7~\2\2\u00b6\u00b7\7~\2\2\u00b7\60\3\2\2\2\u00b8\u00ba")
+        buf.write("\t\2\2\2\u00b9\u00b8\3\2\2\2\u00ba\u00bb\3\2\2\2\u00bb")
+        buf.write("\u00b9\3\2\2\2\u00bb\u00bc\3\2\2\2\u00bc\u00c7\3\2\2\2")
+        buf.write("\u00bd\u00bf\t\3\2\2\u00be\u00bd\3\2\2\2\u00be\u00bf\3")
+        buf.write("\2\2\2\u00bf\u00c1\3\2\2\2\u00c0\u00c2\t\4\2\2\u00c1\u00c0")
+        buf.write("\3\2\2\2\u00c2\u00c3\3\2\2\2\u00c3\u00c1\3\2\2\2\u00c3")
+        buf.write("\u00c4\3\2\2\2\u00c4\u00c6\3\2\2\2\u00c5\u00be\3\2\2\2")
+        buf.write("\u00c6\u00c9\3\2\2\2\u00c7\u00c5\3\2\2\2\u00c7\u00c8\3")
+        buf.write("\2\2\2\u00c8\62\3\2\2\2\u00c9\u00c7\3\2\2\2\u00ca\u00cc")
+        buf.write("\t\5\2\2\u00cb\u00ca\3\2\2\2\u00cc\u00cd\3\2\2\2\u00cd")
+        buf.write("\u00cb\3\2\2\2\u00cd\u00ce\3\2\2\2\u00ce\u00d9\3\2\2\2")
+        buf.write("\u00cf\u00d1\t\3\2\2\u00d0\u00cf\3\2\2\2\u00d0\u00d1\3")
+        buf.write("\2\2\2\u00d1\u00d3\3\2\2\2\u00d2\u00d4\t\4\2\2\u00d3\u00d2")
+        buf.write("\3\2\2\2\u00d4\u00d5\3\2\2\2\u00d5\u00d3\3\2\2\2\u00d5")
+        buf.write("\u00d6\3\2\2\2\u00d6\u00d8\3\2\2\2\u00d7\u00d0\3\2\2\2")
+        buf.write("\u00d8\u00db\3\2\2\2\u00d9\u00d7\3\2\2\2\u00d9\u00da\3")
+        buf.write("\2\2\2\u00da\64\3\2\2\2\u00db\u00d9\3\2\2\2\u00dc\u00dd")
+        buf.write("\5\63\32\2\u00dd\u00de\7*\2\2\u00de\66\3\2\2\2\u00df\u00e0")
+        buf.write("\5\61\31\2\u00e08\3\2\2\2\u00e1\u00e2\5\61\31\2\u00e2")
+        buf.write("\u00e3\7*\2\2\u00e3:\3\2\2\2\u00e4\u00e6\t\5\2\2\u00e5")
+        buf.write("\u00e4\3\2\2\2\u00e6\u00e7\3\2\2\2\u00e7\u00e5\3\2\2\2")
+        buf.write("\u00e7\u00e8\3\2\2\2\u00e8\u00ec\3\2\2\2\u00e9\u00eb\t")
+        buf.write("\5\2\2\u00ea\u00e9\3\2\2\2\u00eb\u00ee\3\2\2\2\u00ec\u00ea")
+        buf.write("\3\2\2\2\u00ec\u00ed\3\2\2\2\u00ed\u00f2\3\2\2\2\u00ee")
+        buf.write("\u00ec\3\2\2\2\u00ef\u00f1\t\6\2\2\u00f0\u00ef\3\2\2\2")
+        buf.write("\u00f1\u00f4\3\2\2\2\u00f2\u00f0\3\2\2\2\u00f2\u00f3\3")
+        buf.write("\2\2\2\u00f3<\3\2\2\2\u00f4\u00f2\3\2\2\2\u00f5\u00f7")
+        buf.write("\t\7\2\2\u00f6\u00f5\3\2\2\2\u00f7\u00f8\3\2\2\2\u00f8")
+        buf.write("\u00f6\3\2\2\2\u00f8\u00f9\3\2\2\2\u00f9\u00fa\3\2\2\2")
+        buf.write("\u00fa\u00fb\b\37\2\2\u00fb>\3\2\2\2\u00fc\u00fd\7$\2")
+        buf.write("\2\u00fd\u0108\t\4\2\2\u00fe\u0100\t\3\2\2\u00ff\u00fe")
+        buf.write("\3\2\2\2\u00ff\u0100\3\2\2\2\u0100\u0102\3\2\2\2\u0101")
+        buf.write("\u0103\t\4\2\2\u0102\u0101\3\2\2\2\u0103\u0104\3\2\2\2")
+        buf.write("\u0104\u0102\3\2\2\2\u0104\u0105\3\2\2\2\u0105\u0107\3")
+        buf.write("\2\2\2\u0106\u00ff\3\2\2\2\u0107\u010a\3\2\2\2\u0108\u0106")
+        buf.write("\3\2\2\2\u0108\u0109\3\2\2\2\u0109\u010b\3\2\2\2\u010a")
+        buf.write("\u0108\3\2\2\2\u010b\u010c\7$\2\2\u010c@\3\2\2\2\u010d")
+        buf.write("\u0115\t\6\2\2\u010e\u0110\t\b\2\2\u010f\u0111\t\6\2\2")
+        buf.write("\u0110\u010f\3\2\2\2\u0111\u0112\3\2\2\2\u0112\u0110\3")
+        buf.write("\2\2\2\u0112\u0113\3\2\2\2\u0113\u0115\3\2\2\2\u0114\u010d")
+        buf.write("\3\2\2\2\u0114\u010e\3\2\2\2\u0115B\3\2\2\2\u0116\u0117")
+        buf.write("\5A!\2\u0117\u0119\7\60\2\2\u0118\u011a\5A!\2\u0119\u0118")
+        buf.write("\3\2\2\2\u0119\u011a\3\2\2\2\u011aD\3\2\2\2\36\2Uajpy")
+        buf.write("\u0081\u008a\u0091\u009a\u00bb\u00be\u00c3\u00c7\u00cd")
+        buf.write("\u00d0\u00d5\u00d9\u00e7\u00ec\u00f2\u00f8\u00ff\u0104")
+        buf.write("\u0108\u0112\u0114\u0119\3\b\2\2")
+        return buf.getvalue()
+
+
+class FormulaLexer(Lexer):
+
+    atn = ATNDeserializer().deserialize(serializedATN())
+
+    decisionsToDFA = [ DFA(ds, i) for i, ds in enumerate(atn.decisionToState) ]
+
+    T__0 = 1
+    T__1 = 2
+    T__2 = 3
+    T__3 = 4
+    TRUE = 5
+    FALSE = 6
+    NOT = 7
+    UNIVERSAL = 8
+    EXISTENTIAL = 9
+    COUNTING = 10
+    AND = 11
+    OR = 12
+    XOR = 13
+    IMPL = 14
+    EQ = 15
+    IN = 16
+    EQUAL = 17
+    UNEQUAL = 18
+    LESS = 19
+    GREATER = 20
+    LESS_EQUAL = 21
+    GREATER_EQUAL = 22
+    WITH = 23
+    PREDICATE = 24
+    VAR = 25
+    FUNCTION = 26
+    FIXED_DOMAIN = 27
+    WS = 28
+    STR_VALUE = 29
+    INT_VALUE = 30
+    FLOAT_VALUE = 31
+
+    channelNames = [ u"DEFAULT_TOKEN_CHANNEL", u"HIDDEN" ]
+
+    modeNames = [ "DEFAULT_MODE" ]
+
+    literalNames = [ "<INVALID>",
+            "'.'", "','", "')'", "'('", "'->'", "'<->'", "'in'", "'='", 
+            "'!='", "'<'", "'>'", "'<='", "'>='", "'||'" ]
+
+    symbolicNames = [ "<INVALID>",
+            "TRUE", "FALSE", "NOT", "UNIVERSAL", "EXISTENTIAL", "COUNTING", 
+            "AND", "OR", "XOR", "IMPL", "EQ", "IN", "EQUAL", "UNEQUAL", 
+            "LESS", "GREATER", "LESS_EQUAL", "GREATER_EQUAL", "WITH", "PREDICATE", 
+            "VAR", "FUNCTION", "FIXED_DOMAIN", "WS", "STR_VALUE", "INT_VALUE", 
+            "FLOAT_VALUE" ]
+
+    ruleNames = [ "T__0", "T__1", "T__2", "T__3", "TRUE", "FALSE", "NOT", 
+                  "UNIVERSAL", "EXISTENTIAL", "COUNTING", "AND", "OR", "XOR", 
+                  "IMPL", "EQ", "IN", "EQUAL", "UNEQUAL", "LESS", "GREATER", 
+                  "LESS_EQUAL", "GREATER_EQUAL", "WITH", "LOWER_ALPHA_UNDERSCORE", 
+                  "UPPER_ALPHA_UNDERSCORE", "PREDICATE", "VAR", "FUNCTION", 
+                  "FIXED_DOMAIN", "WS", "STR_VALUE", "INT_VALUE", "FLOAT_VALUE" ]
+
+    grammarFileName = "Formula.g4"
+
+    def __init__(self, input=None, output:TextIO = sys.stdout):
+        super().__init__(input, output)
+        self.checkVersion("4.9.3")
+        self._interp = LexerATNSimulator(self, self.atn, self.decisionsToDFA, PredictionContextCache())
+        self._actions = None
+        self._predicates = None
+
+